refactor: enum 패키지를 사용하여 enum과 entity를 구분
