--- src/main/java/com/safevillage/safevillage/domain/auth/entity/Role.java
@@ -1,7 +1,0 @@
-package com.safevillage.safevillage.domain.auth.entity;
-
-public enum Role {
-
-    USER,
-    ADMIN
-}
