+++ src/main/java/com/safevillage/safevillage/domain/auth/enums/Role.java
@@ -1,0 +1,7 @@
+package com.safevillage.safevillage.domain.auth.enums;
+
+public enum Role {
+
+    USER,
+    ADMIN
+}
