+++ src/main/java/com/safevillage/safevillage/domain/reports/dto/ReportResponse.java
@@ -1,8 +1,8 @@
 package com.safevillage.safevillage.domain.reports.dto;
 
-import com.safevillage.safevillage.domain.reports.entity.DangerLevel;
-import com.safevillage.safevillage.domain.reports.entity.ReportCategory;
-import com.safevillage.safevillage.domain.reports.entity.ReportStatus;
+import com.safevillage.safevillage.domain.reports.enums.DangerLevel;
+import com.safevillage.safevillage.domain.reports.enums.ReportCategory;
+import com.safevillage.safevillage.domain.reports.enums.ReportStatus;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
