+++ src/main/java/com/safevillage/safevillage/domain/reports/service/ReportService.java
@@ -5,10 +5,10 @@ import com.safevillage.safevillage.domain.auth.repository.UserRepository;
 import com.safevillage.safevillage.domain.reports.dto.ReportCreateRequest;
 import com.safevillage.safevillage.domain.reports.dto.ReportLikeResponse;
 import com.safevillage.safevillage.domain.reports.dto.ReportResponse;
-import com.safevillage.safevillage.domain.reports.entity.DangerLevel;
+import com.safevillage.safevillage.domain.reports.enums.DangerLevel;
 import com.safevillage.safevillage.domain.reports.entity.Report;
 import com.safevillage.safevillage.domain.reports.entity.ReportLike;
-import com.safevillage.safevillage.domain.reports.entity.ReportStatus;
+import com.safevillage.safevillage.domain.reports.enums.ReportStatus;
 import com.safevillage.safevillage.domain.reports.repository.ReportLikeRepository;
 import com.safevillage.safevillage.domain.reports.repository.ReportRepository;
 import org.springframework.data.domain.Pageable;
